requirements.txt
mcp[server]
fastmcp
fastapi
httpx
python-dotenv
